Bug 1897786 - Part 3: Support wasm target with minimal code and dependencies
