+++ tools/src/cmd_pipeline/symbol_graph.rs
@@ -25,6 +25,8 @@ use crate::{
     },
 };
 
+pub use crate::symbol_graph_edge_kind::EdgeKind;
+
 use super::{
     cmd_graph::{GraphHierarchy, GraphLayout},
     interface::OverloadInfo,
@@ -2000,19 +2002,6 @@ pub struct SymbolGraphNodeSet {
 pub struct SymbolGraphEdgeSet {
     pub edge_infos: Vec<EdgeInfo>,
     edge_lookup: HashMap<(u32, u32), u32>,
-}
-
-#[derive(Clone)]
-pub enum EdgeKind {
-    Default, // solid line, closed arrow ("normal")
-    // These value are meant to be UML-ish
-    Inheritance,    // solid line, open arrow ("onormal")
-    Implementation, // dashed line, open arrow ("onormal")
-    Composition,    // solid line, closed diamond ("diamond")
-    Aggregation,    // solid line, open diamond ("odiamond")
-    // These are more specific searchfox concepts
-    IPC,           // dotted line, weird vee arrow ("vee")
-    CrossLanguage, // JNI-like; solid line, left-half-closed arrow ("lnormal")
 }
 
 #[derive(Clone, PartialEq)]
